fix windows scripts

+++ install.ps1
@@ -1,10 +1,16 @@
-pushd /
+pushd "$HOME"
 
-irm get.scoop.sh -outfile 'scoop.ps1'
-.\scoop.ps1 -RunAsAdmin | out-null
-rm scoop.ps1
+if ((Get-Command scoop -ErrorAction SilentlyContinue) -eq $null) {
+    irm get.scoop.sh -outfile 'scoop.ps1'
+    .\scoop.ps1 -RunAsAdmin | out-null
+    rm scoop.ps1
+}
 
-scoop install vscode
+scoop bucket add nerd-fonts
+scoop bucket add extras
+scoop install firacode vscode sd
+reg import "$HOME\scoop\apps\vscode\current\install-github-integration.reg"
+
 code --install-extension yathink3.carbon-react-color-theme --force
 code --install-extension pkief.material-icon-theme --force
 code --install-extension esbenp.prettier-vscode --force
@@ -12,17 +18,19 @@ code --install-extension be5invis.vscode-custom-css --force
 code --install-extension apility.beautify-blade --force
 code --install-extension isudox.vscode-jetbrains-keybindings --force
 code --install-extension donjayamanne.githistory --force
+code --install-extension RimuruChan.vscode-fix-checksums-next --force
 
-scoop bucket add nerd-fonts
-scoop install firacode
+$repo = "https://raw.githubusercontent.com/chrishenn/vscode/refs/heads/main"
+$cfg = "$HOME/scoop/apps/vscode/current/data/user-data/User"
 
-if (! (test-path /vscode)) {
-    git clone git@github.com:chrishenn/vscode.git
-    cd /vscode
-} else {
-    cd /vscode
-    git pull
-}
-cp /vscode/settings.json "$HOME\scoop\apps\vscode\current\data\user-data\User\settings.json" -force
-cp /vscode/keybindings.json "$HOME\scoop\apps\vscode\current\data\user-data\User\keybindings.json" -force
+[void](mkdir -ea 0 "$cfg")
+
+curl "$repo/settings.json" -o "$cfg/settings.json"
+curl "$repo/keybindings.json" -o "$cfg/keybindings.json"
+curl "$repo/code.css" -o "$cfg/code.css"
+curl "$repo/code.js" -o "$cfg/code.js"
+
+sd 'CODE_CSS' "$cfg/code.css".Replace('\', '/') "$cfg/settings.json"
+sd 'CODE_JS' "$cfg/code.js".Replace('\', '/') "$cfg/settings.json"
+
 popd
